Update README.md
Clear up some confusion in the docs. It's not optional to have more than 1 item defined because it simply wont work. You can't have more than 1 method of calling the command.

setting a config 
    array(
        'prefix' => '!',
        'nick' => false
    )

would not work as expected since the config is checking for isset on the nick and not checking the other rules.